+++ src/main/java/JuDBu/custommaster/controller/RootController.java
@@ -1,4 +1,4 @@
-package JuDBu.custommaster;
+package JuDBu.custommaster.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
