--- src/main/java/JuDBu/custommaster/entity/ord/request/OrdRepository.java
@@ -1,8 +1,0 @@
-package JuDBu.custommaster.entity.ord.request;
-
-import JuDBu.custommaster.entity.ord.Ord;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrdRepository extends JpaRepository<Ord, Long> {
-
-}
