+++ src/main/resources/templates/ord/ord-detail.html
@@ -6,5 +6,6 @@
 </head>
 <body>
     <h2>주문 상세</h2>
+    <hr>
 </body>
 </html>
